+++ src/app/pages/entity/application/application.component.spec.ts
@@ -1,0 +1,32 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ApplicationComponent } from './application.component';
+import { AppModule } from 'src/app/app.module';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+
+describe('ApplicationComponent', () => {
+  let component: ApplicationComponent;
+  let fixture: ComponentFixture<ApplicationComponent>;
+
+  afterEach(() => {
+    TestBed.resetTestingModule();
+  });
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [AppModule, HttpClientTestingModule],
+    }).compileComponents();
+  });
+
+  afterEach(() => {
+    TestBed.resetTestingModule();
+  });
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ApplicationComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});