+++ src/app/pages/public/recover/recover.component.spec.ts
@@ -1,0 +1,27 @@
+import { Title } from '@angular/platform-browser';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { AppModule } from '../../../app.module';
+import { RecoverComponent } from './recover.component';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+
+describe('RecoverComponent', () => {
+  let component: RecoverComponent;
+  let fixture: ComponentFixture<RecoverComponent>;
+
+  afterEach(() => {
+    TestBed.resetTestingModule();
+  });
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [AppModule, HttpClientTestingModule],
+      providers: [{ provide: Title, useClass: Title }],
+    }).compileComponents();
+  });
+
+  it('should create', () => {
+    fixture = TestBed.createComponent(RecoverComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+    expect(component).toBeTruthy();
+  });
+});