+++ src/app/routes.ts
@@ -1,20 +1,21 @@
-import { ForgotPassComponent } from './pages/forgot-pass/forgot-pass.component';
+import { ForgotPassComponent } from './pages/public/forgot-pass/forgot-pass.component';
 import { DashboardComponent } from './pages/dashboard/dashboard.component';
 import { Routes } from '@angular/router';
-import { LoginComponent } from './pages/login/login.component';
-import { UsersComponent } from './pages/users/users.component';
+import { LoginComponent } from './pages/public/login/login.component';
+import { UsersComponent } from './pages/lists/users/users.component';
 import { ProfileComponent } from './pages/profile/profile.component';
-import { UserComponent } from './pages/user/user.component';
-import { SignupComponent } from './pages/signup/signup.component';
-import { AccountsComponent } from './pages/accounts/accounts.component';
-import { PermissionsComponent } from './pages/permissions/permissions.component';
-import { ApplicationsComponent } from './pages/applications/applications.component';
-import { AccountComponent } from './pages/account/account.component';
-import { ApplicationComponent } from './pages/application/application.component';
-import { VerifyComponent } from './pages/verify/verify.component';
-import { RecoverComponent } from './pages/recover/recover.component';
+import { UserComponent } from './pages/entity/user/user.component';
+import { SignupComponent } from './pages/public/signup/signup.component';
+import { AccountsComponent } from './pages/lists/accounts/accounts.component';
+import { PermissionsComponent } from './pages/lists/permissions/permissions.component';
+import { ApplicationsComponent } from './pages/lists/applications/applications.component';
+import { AccountComponent } from './pages/entity/account/account.component';
+import { ApplicationComponent } from './pages/entity/application/application.component';
+import { VerifyComponent } from './pages/public/verify/verify.component';
+import { RecoverComponent } from './pages/public/recover/recover.component';
 
 export const routes: Routes = [
+  // public
   { path: '', redirectTo: '/login', pathMatch: 'full' },
   {
     path: 'login',
@@ -41,49 +42,84 @@ export const routes: Routes = [
     component: ForgotPassComponent,
     canActivate: ForgotPassComponent.guards,
   },
+  // Private
   {
     path: 'dashboard',
     component: DashboardComponent,
     canActivate: DashboardComponent.guards,
+    data: {
+      breadcrumb: 'DASBHOARD',
+    },
   },
   {
     path: 'accounts',
     component: AccountsComponent,
     canActivate: AccountsComponent.guards,
-  },
-  {
-    path: 'accounts/:id',
-    component: AccountComponent,
-    canActivate: AccountComponent.guards,
+    data: {
+      breadcrumb: 'ACCOUNTS',
+    },
+
+    children: [
+      {
+        path: ':id',
+        component: AccountComponent,
+        canActivate: AccountComponent.guards,
+        data: {
+          breadcrumb: 'ACCOUNT',
+        },
+      },
+    ],
   },
   {
     path: 'permissions',
     component: PermissionsComponent,
     canActivate: PermissionsComponent.guards,
+    data: {
+      breadcrumb: 'PERMISSIONS',
+    },
   },
   {
     path: 'applications',
     component: ApplicationsComponent,
     canActivate: ApplicationsComponent.guards,
-  },
-  {
-    path: 'applications/:id',
-    component: ApplicationComponent,
-    canActivate: ApplicationComponent.guards,
+    data: {
+      breadcrumb: 'APPLICATIONS',
+    },
+    children: [
+      {
+        path: ':id',
+        component: ApplicationComponent,
+        canActivate: ApplicationComponent.guards,
+        data: {
+          breadcrumb: 'APPLICATION',
+        },
+      },
+    ],
   },
   {
     path: 'users',
     component: UsersComponent,
     canActivate: UsersComponent.guards,
-  },
-  {
-    path: 'users/:id',
-    component: UserComponent,
-    canActivate: UserComponent.guards,
+    data: {
+      breadcrumb: 'USERS',
+    },
+    children: [
+      {
+        path: 'users/:id',
+        component: UserComponent,
+        canActivate: UserComponent.guards,
+        data: {
+          breadcrumb: 'USER',
+        },
+      },
+    ],
   },
   {
     path: 'profile',
     component: ProfileComponent,
     canActivate: ProfileComponent.guards,
+    data: {
+      breadcrumb: 'PROFILE',
+    },
   },
 ];