+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/QbitArtifacts/caste-panel/compare/v1.1.0...v1.2.0) (2021-03-16)
+
+
+### Features
+
+* **application:** added some information to application details page | [#12](https://github.com/QbitArtifacts/caste-panel/issues/12) ([e01c0d7](https://github.com/QbitArtifacts/caste-panel/commit/e01c0d713586700c92015fd28878ca13f2591502))
+* **application:** added some information to application details page | [#12](https://github.com/QbitArtifacts/caste-panel/issues/12) ([4280559](https://github.com/QbitArtifacts/caste-panel/commit/42805597207870078cdc32f78e764b0e4adc4792))
+
+
+### Bug Fixes
+
+* **accounts:** create account failed | fixes [#23](https://github.com/QbitArtifacts/caste-panel/issues/23) ([3418e01](https://github.com/QbitArtifacts/caste-panel/commit/3418e013e336abc8aab1225a56411396922d09b5))
+* **apps:** added realm to app list | fixes [#7](https://github.com/QbitArtifacts/caste-panel/issues/7) ([e7ce8ab](https://github.com/QbitArtifacts/caste-panel/commit/e7ce8ab290a57c499e55b1c90b3afa139c4a6682))
+* **deps:** updated caste-client-ng with updated entities ([3e4c755](https://github.com/QbitArtifacts/caste-panel/commit/3e4c755595be2297a0a250aa7080ab44e7524675))
+* **deps:** Upgraded angular from 8 to 11 ([b43904e](https://github.com/QbitArtifacts/caste-panel/commit/b43904ec78fdbd60476394056c8b58c9a3b23c98))
+* **i18n:** added missing translations | fixes [#24](https://github.com/QbitArtifacts/caste-panel/issues/24) ([0e94502](https://github.com/QbitArtifacts/caste-panel/commit/0e9450246892c4a97760884a282cfeeac5a7ed69))
+* **recover:** not sending realm in recover, updated deps | fixes [#11](https://github.com/QbitArtifacts/caste-panel/issues/11) ([707c05f](https://github.com/QbitArtifacts/caste-panel/commit/707c05fc0a811956235dfbfe4042e5a802c56473))
+* **tests:** added some tests, and added HttpClientTestingModule to all tests ([68de34b](https://github.com/QbitArtifacts/caste-panel/commit/68de34bbf9a622305d0c0844448c02ae7cdb2110))
+* **tests:** fixed failing tests ([2e7d329](https://github.com/QbitArtifacts/caste-panel/commit/2e7d32984726232a36b1f624b6d56032ef4a7ad2))
+* **tests:** fixed some failing tests ([24d4fe2](https://github.com/QbitArtifacts/caste-panel/commit/24d4fe2b7674e34a6fe364f481ea11de0d997f76))
+* **ui:** fixed some trivial ui things ([5b406a6](https://github.com/QbitArtifacts/caste-panel/commit/5b406a689d16384699475deee32b16eb93c6e4b2))
+* **user:** fixed user details page not loading ([a05882f](https://github.com/QbitArtifacts/caste-panel/commit/a05882f10a90d7d71a20277ca5ef563eb939c612))
+* **version:** moved version to sidemenu and added api version ([bbf6dcd](https://github.com/QbitArtifacts/caste-panel/commit/bbf6dcd34fd7e693946090d22b7a7b529e99f057))
+* fixed missing imports ([00a1f5a](https://github.com/QbitArtifacts/caste-panel/commit/00a1f5af01fd8013d71276e62ed6aafaa6553040))
+* load of fixes, improving ui, more... ([418326b](https://github.com/QbitArtifacts/caste-panel/commit/418326b3b59f5a552bdc3ffe745e8bfb4b086221))
+
 ## [1.1.0](https://github.com/QbitArtifacts/caste-panel/compare/v1.0.1...v1.1.0) (2020-11-30)